Extracting test common functions, coderefactor
